Refactor useChat to handle function calls in streamed responses
This commit refactors the `useChat` function to handle function calls in streamed responses. It adds a new parameter, `onFunctionCall`, which is called when a function call is received from the API. The user can define their own handler for this event and return a new chat request object that will be sent back to the API. If no response is returned, the loop will break and append any remaining messages to the chat history. This change also updates some types and renames `RequestOptions` to `ChatRequestOptions`.